requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
mangum==0.11.0
pynamodb==5.1.0
pytest==6.2.5
pytest-env==0.6.2
pytest-mock==3.6.1
sentry-sdk==1.5.2
argon2-cffi==21.3.0 
PyJWT==2.3.0
boto3==1.20.41
requests==2.27.1
requests-mock==1.9.3 
freezegun==1.1.0
python-dotenv==0.19.2